+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 1.0.1
 
 - Add support for a lot more promise implementations
+- Add `beforeEach` and `afterEach` that accept async functions
 
 ## 1.0.0